Fix the logic to support C++17 compiler and remove C++11

+++ jenkins-scripts/docker/lib/docker_generate_dockerfile.bash
@@ -306,12 +306,15 @@ DELIM_DOCKER31
 # Beware of moving this code since it needs to run update-alternative after
 # installing the default compiler in PACKAGES_CACHE_AND_CHECK_UPDATES
 if ${NEED_C17_COMPILER}; then
+# Newer distributions don't need custom compiler
+if ${DISTRO} == 'bionic' || ${DISTRO} == 'buster'; then
 cat >> Dockerfile << DELIM_GCC8
    RUN apt-get update \\
    && apt-get install -y g++-8 \\
    && rm -rf /var/lib/apt/lists/* \\
    && update-alternatives --install /usr/bin/gcc gcc /usr/bin/gcc-8 800 --slave /usr/bin/g++ g++ /usr/bin/g++-8 --slave /usr/bin/gcov gcov /usr/bin/gcov-8
 DELIM_GCC8
+fi
 fi
 
 if ${USE_SQUID}; then
